requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11
webrtcvad>=2.0.10

# Computer Vision
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0
Pillow>=10.0.0

# NLP and Text Processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
jieba>=0.42.0

# Data Processing and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0

# Location and Sensor Data
geopy>=2.3.0
folium>=0.14.0
accelerometer>=0.1.0

# Database and Storage
pymongo>=4.4.0
redis>=4.6.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
websockets>=11.0.0

# Security and Encryption
cryptography>=41.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Mobile and Edge Computing
onnx>=1.14.0
onnxruntime>=1.15.0

# Real-time Processing
aiohttp>=3.8.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0 